requirements.txt
Django
Jinja2

aiohttp
beautifulsoup4
more-itertools

python-dateutil
pytz
